requirements.txt
flake8
Flask
graphql-core-next
psycopg2
pytest
python-dotenv
requests